refactor: move BloomNavHost and MainScaffold into navigation package and clean MainActivity
