+++ app/src/main/java/com/example/bloom/ui/screens/settings/SettingsScreen.kt
@@ -3,7 +3,6 @@ package com.example.bloom.ui.screens.settings
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
@@ -11,11 +10,13 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 
 @Composable
-fun SettingsScreen() {
-    var isDarkMode by rememberSaveable { mutableStateOf(false) }
-
+fun SettingsScreen(
+    isDarkMode: Boolean,
+    onThemeToggle: (Boolean) -> Unit,
+    modifier: Modifier = Modifier
+) {
     Column(
-        modifier = Modifier
+        modifier = modifier
             .fillMaxSize()
             .padding(24.dp),
         verticalArrangement = Arrangement.spacedBy(24.dp),
@@ -36,18 +37,13 @@ fun SettingsScreen() {
             horizontalArrangement = Arrangement.SpaceBetween,
             verticalAlignment = Alignment.CenterVertically
         ) {
-            Text(
-                text = "Dark Mode",
-                style = MaterialTheme.typography.bodyLarge
-            )
-
+            Text("Dark Mode", style = MaterialTheme.typography.bodyLarge)
             Switch(
                 checked = isDarkMode,
-                onCheckedChange = { isDarkMode = it }
+                onCheckedChange = { onThemeToggle(it) }
             )
         }
 
-        // Optional: Preview of theme
         Text(
             text = if (isDarkMode) "🌙 Dark Mode Enabled" else "☀️ Light Mode Enabled",
             style = MaterialTheme.typography.bodyMedium
